Refactor file structure: move `IterableInbox` to `inbox` directory and update imports; create new hooks index file
