+++ src/index.tsx
@@ -32,9 +32,9 @@ export {
 } from './IterableInAppClasses';
 export { IterableInAppManager } from './IterableInAppManager';
 export { IterableInAppMessage } from './IterableInAppMessage';
-export { IterableInbox, type IterableInboxProps } from './IterableInbox';
+export { IterableInbox, type IterableInboxProps } from './inbox/IterableInbox';
 export type { IterableInboxCustomizations } from './IterableInboxCustomizations';
 export { IterableInboxEmptyState } from './IterableInboxEmptyState';
 export { IterableInboxMessageCell } from './IterableInboxMessageCell';
-export { useAppStateListener } from './useAppStateListener';
-export { useDeviceOrientation } from './useDeviceOrientation';
+export { useAppStateListener } from './hooks/useAppStateListener';
+export { useDeviceOrientation } from './hooks/useDeviceOrientation';
